Move more classes to dagger
PluginTracker, PluginTrackerImpl, PluginDependency and
PluginDependencyProvider is now Dagger instead of Dependency.

Issue: #16
