--- pluginlib/src/main/kotlin/com/prefect47/pluginlib/impl/di/AppContext.kt
@@ -1,7 +1,0 @@
-package com.prefect47.pluginlib.impl.di
-
-import javax.inject.Qualifier
-
-@Qualifier
-@kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
-annotation class AppContext
