+++ pluginlib/src/main/kotlin/com/prefect47/pluginlib/impl/di/PluginLibraryDI.kt
@@ -12,20 +12,16 @@
  * permissions and limitations under the License.
  */
 
-package com.prefect47.pluginlib
+package com.prefect47.pluginlib.impl.di
 
 import android.content.Context
-import com.prefect47.pluginlib.impl.Dependency
-import com.prefect47.pluginlib.impl.di.PluginLibraryDI
-import com.prefect47.pluginlib.plugin.PluginLibraryControl
+import com.prefect47.pluginlib.impl.di.component.DaggerPluginLibraryComponent
+import com.prefect47.pluginlib.impl.di.component.PluginLibraryComponent
 
-object PluginLibrary {
-    const val ARG_CLASSNAME = "pluginClassName"
+internal object PluginLibraryDI {
+    lateinit var component: PluginLibraryComponent
 
-    fun init(context: Context) {
-        PluginLibraryDI.init(context)
-        Dependency.init(context)
+    internal fun init(context: Context) {
+        component = DaggerPluginLibraryComponent.builder().context(context).build()
     }
-
-    fun getControl() = Dependency[PluginLibraryControl::class]
 }
