+++ pluginlib/src/main/kotlin/com/prefect47/pluginlib/impl/di/module/PluginLibraryFactoryModule.kt
@@ -1,6 +1,7 @@
 package com.prefect47.pluginlib.impl.di.module
 
 import com.prefect47.pluginlib.impl.*
+import com.prefect47.pluginlib.plugin.PluginDependency
 import dagger.Binds
 import dagger.Module
 import javax.inject.Singleton
@@ -18,4 +19,8 @@ abstract class PluginLibraryFactoryModule {
     @Singleton
     @Binds
     abstract fun bindPluginTrackerFactory(factory: PluginTrackerImpl.Factory): PluginTrackerFactory
+
+    @Singleton
+    @Binds
+    abstract fun bindDependencyProvider(provider: PluginDependencyProvider): PluginDependency.DependencyProvider
 }
