+++ pluginlib/src/main/kotlin/com/prefect47/pluginlib/plugin/PluginDependency.kt
@@ -15,6 +15,7 @@
 
 package com.prefect47.pluginlib.plugin
 
+import com.prefect47.pluginlib.impl.di.PluginLibraryDI
 import com.prefect47.pluginlib.plugin.annotations.ProvidesInterface
 import kotlin.reflect.KClass
 
@@ -22,10 +23,10 @@ import kotlin.reflect.KClass
 object PluginDependency {
     const val VERSION = 1
 
-    lateinit var sProvider: DependencyProvider
+    private val provider: DependencyProvider by lazy { PluginLibraryDI.component.getDependencyProvider() }
 
     operator fun <T: Any> get(p: Plugin, cls: KClass<T>): T {
-        return sProvider[p, cls]
+        return provider[p, cls]
     }
 
     abstract class DependencyProvider {
